oops

+++ sys/stand/boot/boot.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: boot_loadfile.c,v 1.1 2003/04/17 03:53:53 drahn Exp $	*/
+/*	$OpenBSD: boot.c,v 1.24 2003/04/17 12:06:05 mickey Exp $	*/
 
 /*
  * Copyright (c) 2003 Dale Rahn
@@ -81,7 +81,7 @@ boot(bootdev)
 		st = 0;
 
 		printf("booting %s: ", cmd.path);
-		marks[MARK_START] = cmd.addr;
+		marks[MARK_START] = (u_long)cmd.addr;
 		if (loadfile(cmd.path, marks, LOAD_ALL) >= 0)
 			break;
